# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SCPM API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/solidCP/vps/{solidcp_item_id}/provisioning-status": {
            "get": {
                "tags": [
                    "VPS Provisioning Status"
                ],
                "description": "Get VPS Provisioning Status. Whether the server is prepared or not. Should be used after calling All in One User/Package/Vps",
                "operationId": "177545e5a04cc9b11f6be5bc055a99bb",
                "parameters": [
                    {
                        "name": "solidcp_item_id",
                        "in": "path",
                        "description": "the id that provides the VPS creation method",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "id_enterprise_dispatcher",
                        "in": "query",
                        "description": "if not selected, the default is used. No need to choose if only one enterprise is used",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "provisioning_status": {
                                            "description": "OK, Error, InProgress",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solidCP/vps/{solidcp_item_id}/state": {
            "get": {
                "tags": [
                    "VPS State of fully installed VM"
                ],
                "description": "Get state of VM Running/Stopped/ets. Must be used only if VPS Provisioning Status shows - OK",
                "operationId": "8e4d187e6bf4f8b2f4ef489045332896",
                "parameters": [
                    {
                        "name": "solidcp_item_id",
                        "in": "path",
                        "description": "the id that provides the VPS creation method",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "id_enterprise_dispatcher",
                        "in": "query",
                        "description": "if not selected, the default is used. No need to choose if only one enterprise is used",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "state": {
                                            "description": "After installation, we need to get the state = Running, everything else in this case means a problem. If state = Starting, then need to wait.",
                                            "type": "string",
                                            "example": "Snapshotting or Migrating or Deleted or Unknown or Other or Running or Off or Stopping or Saved or Paused or Starting or Reset or Saving or Pausing or Resuming or FastSaved or FastSaving or RunningCritical or OffCritical or StoppingCritical or SavedCritical or PausedCritical or StartingCritical or ResetCritical or SavingCritical or PausingCritical or ResumingCritical or FastSavedCritical or FastSavingCritical"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solidCP/vps/{vps_ip_address}/status": {
            "put": {
                "tags": [
                    "Change VPS package status over IPv4"
                ],
                "description": "Change VPS and its packet status over IPv4",
                "operationId": "c2b983dae2e8e2df8418895803e497a1",
                "parameters": [
                    {
                        "name": "vps_ip_address",
                        "in": "path",
                        "description": "Main IPv4 address",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ipv4"
                        }
                    },
                    {
                        "name": "id_enterprise_dispatcher",
                        "in": "query",
                        "description": "if not selected, the default is used. No need to choose if only one enterprise is used",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vps_status"
                                ],
                                "properties": {
                                    "vps_status": {
                                        "type": "string",
                                        "enum": [
                                            "Active",
                                            "Suspended",
                                            "Cancelled"
                                        ]
                                    },
                                    "id_enterprise_dispatcher": {
                                        "description": "if not selected, then use the default one",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solidCP/vps/available-spaces/count": {
            "get": {
                "tags": [
                    "Number of Available VPS Spaces aka VPS storages"
                ],
                "description": "Provides the number of available spaces for a specific package",
                "operationId": "b8bb977e8b74d9633cc424a09beb618b",
                "parameters": [
                    {
                        "name": "server_package_name",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Silver RDP23"
                    },
                    {
                        "name": "server_location_name",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Meppel, Netherlands"
                    },
                    {
                        "name": "server_ip_amount",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_enterprise_dispatcher",
                        "in": "query",
                        "description": "if not selected, the default is used. No need to choose if only one enterprise is used",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ignore_node_ids[]",
                        "in": "query",
                        "description": "nodes ids that need to ignore for installation. For example, if the old installation is not yet complete",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "ignore_hosting_space_ids[]",
                        "in": "query",
                        "description": "hostingspace ids (storages) that need to ignore for installation. For example, if the old installation is not yet complete",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "number_available_spaces": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SimpleError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorModel": {
                "properties": {
                    "error": {
                        "nullable": true,
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "violations": {
                        "nullable": true,
                        "type": "array",
                        "items": {
                            "properties": {
                                "propertyPath": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SimpleError": {
                "properties": {
                    "error": {
                        "nullable": true,
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}